Refactor Abort controller to fetch utils
