+++ extension/src/content-script/osu/content.ts
@@ -8,20 +8,10 @@ import { updateFlagsProfile } from "./pages/profile";
 import { updateFlagsTopics } from "./pages/topics";
 import { updateLanguageToOsuLanguage } from "./osuLanguage";
 import { updateFlagsSearch } from "./pages/search";
+import { nextAbortControllerSignal } from "@src/utils/fetchUtils";
 
 const flagClass = "flag-country";
 initConfigure(flagClass);
-let currentAbortController = new AbortController();
-
-export const currentSignal = () => currentAbortController.signal;
-
-export const nextAbortControllerSignal = () => {
-    currentAbortController.abort();
-    currentAbortController = new AbortController();
-
-    return currentAbortController.signal;
-};
-
 export const idFromProfileUrl = (url: string) => {
     return url.split("/")[4];
 };
