+++ extension/src/content-script/osu/pages/search.ts
@@ -1,5 +1,6 @@
 import { TFlagItems, addFlagUsers } from "@src/content-script/osu/flagHtml";
-import { idFromProfileUrl, nextAbortControllerSignal } from "../content";
+import { idFromProfileUrl } from "../content";
+import { nextAbortControllerSignal } from "@src/utils/fetchUtils";
 
 // https://osu.ppy.sh/home/search?mode=user&query=Deif&page=1
 const initPageObserver = new MutationObserver(() => {
