+++ extension/src/content-script/osu/pages/topics.ts
@@ -1,7 +1,7 @@
 // https://osu.ppy.sh/community/forums/topics/1686524?n=3
 
 import { TFlagItems, addFlagUsers as addFlagUsers } from "@src/content-script/osu/flagHtml";
-import { nextAbortControllerSignal } from "../content";
+import { nextAbortControllerSignal } from "@src/utils/fetchUtils";
 
 export const updateFlagsTopics = async () => {
     const url = location.href;
